feat:#66 카드 모임 컴포넌트 생성
